refactor(cspcl): restructure modules, rename async_api to asynchronous
- Rename the async module async_api -> asynchronous and drop the Async
  type prefixes (cspcl::asynchronous::{Cspcl, Sender, Receiver}).
- Split Sender/Receiver out of bundle.rs into a new io.rs; bundle.rs now
  holds only the received-bundle data types.
- Merge address.rs and peer.rs into addressing.rs.
- Update lib.rs re-exports, the renamed integration test, and README.

No behavior changes; pure reorganization. Verified: workspace build,
23 tests pass, cargo doc clean.
